+++ hootsweet/__init__.py
@@ -1,5 +1,5 @@
 from hootsweet.api import HootSweet
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __all__ = ["HootSweet"]