+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### Future
+
+### 0.4.0
+
+- Added /messages POST endpoint
+- Added /messages/{id}/approve
+- Added /messages GET
+- Added /messages DELETE
+
 ### 0.3.0
 
 - Added OAuth2 workflow for authorization